fix: 修复 Telegram Bot 通信（messageId 溢出导致无法接收指令）
- Update.messageId 类型 Int → Long（防止 update_id 溢出）
- 移除 updateId.toInt() 强制转换
- 修复 getUpdates 解析后 lastOffset 错乱问题

+++ app/src/main/java/com/guardeye/TelegramBot.kt
@@ -104,7 +104,7 @@ object TelegramBot {
                     val text = msg.optString("text", "")
                     val chatObj = msg.optJSONObject("chat")
                     val chatIdFrom = chatObj?.optLong("id", 0L)?.toString() ?: ""
-                    updates.add(Update(text, chatIdFrom, updateId.toInt()))
+                    updates.add(Update(text, chatIdFrom, updateId))
                 }
                 updates
             }
@@ -140,5 +140,5 @@ object TelegramBot {
         }
     }
 
-    data class Update(val text: String, val chatId: String, val messageId: Int)
+    data class Update(val text: String, val chatId: String, val messageId: Long)
 }
